LoginView auf Panel umgebaut
