+++ src/controller/common/MainController.java
@@ -1,5 +1,7 @@
 package controller.common;
 
+import model.travel.User;
+import view.common.LoginView;
 import view.common.TripPlannerMain;
 
 import java.awt.event.ActionEvent;
@@ -8,11 +10,11 @@ import java.awt.event.ActionListener;
 public class MainController implements ActionListener {
 
     TripPlannerMain tripPlannerMain;
+    User user;
 
     public MainController(TripPlannerMain tripPlannerMain) {
         this.tripPlannerMain = tripPlannerMain;
     }
-
 
     @Override
     public void actionPerformed(ActionEvent e) {
@@ -27,4 +29,12 @@ public class MainController implements ActionListener {
         }
     }
 
+    public void openLogin() {
+        tripPlannerMain.addView("Login", new LoginView(this));
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+        tripPlannerMain.setUsername(user.getUsername());
+    }
 }
