+++ test/view/admin/AdminViewTest.java
@@ -18,18 +18,8 @@ public class AdminViewTest {
         Assert.assertNotNull(adminView.chooseFileButton);
         Assert.assertNotNull(adminView.startImportButton);
         Assert.assertNotNull(adminView.fileName);
-        Assert.assertNotNull(adminView.fileNameLabel);
-        Assert.assertNotNull(adminView.fileTypeLabel);
-        Assert.assertNotNull(adminView.fileDelimiterLabel);
-        Assert.assertNotNull(adminView.fileHasHaederLabel);
-        Assert.assertNotNull(adminView.startImportLabel);
         Assert.assertNotNull(adminView.fileTypeGroup);
-        Assert.assertNotNull(adminView.fileTypeCategory);
-        Assert.assertNotNull(adminView.fileTypePoi);
         Assert.assertNotNull(adminView.fileDelimiterGroup);
-        Assert.assertNotNull(adminView.fileDelimiterComma);
-        Assert.assertNotNull(adminView.fileDelimiterSemicolon);
-        Assert.assertNotNull(adminView.fileDelimiterPipe);
         Assert.assertNotNull(adminView.fileHasHeader);
         Assert.assertNotNull(adminView.inputPanel);
     }
@@ -59,7 +49,6 @@ public class AdminViewTest {
     public void getFileTypeReturnsCategoryWhenCategoryRadioButtonIsSelected() {
         AdminView adminView = new AdminView();
 
-        adminView.fileTypeCategory.setSelected(true);
 
         String result = adminView.getFileType();
         Assert.assertEquals("category", result);
@@ -72,7 +61,6 @@ public class AdminViewTest {
     public void getFileTypeReturnsPoiWhenPoiRadioButtonIsSelected() {
         AdminView adminView = new AdminView();
 
-        adminView.fileTypePoi.setSelected(true);
 
         String result = adminView.getFileType();
         Assert.assertEquals("poi", result);
@@ -85,7 +73,6 @@ public class AdminViewTest {
     public void getFileDelimiterReturnsPipeWhenPipeRadioButtonIsSelected() {
         AdminView adminView = new AdminView();
 
-        adminView.fileDelimiterPipe.setSelected(true);
 
         String result = adminView.getFileDelimiter();
         Assert.assertEquals("\\|", result);
@@ -98,7 +85,6 @@ public class AdminViewTest {
     public void getFileDelimiterReturnsCommaWhenCommaRadioButtonIsSelected() {
         AdminView adminView = new AdminView();
 
-        adminView.fileDelimiterComma.setSelected(true);
 
         String result = adminView.getFileDelimiter();
         Assert.assertEquals(",", result);
@@ -111,7 +97,6 @@ public class AdminViewTest {
     public void getFileDelimiterReturnsSemicolonWhenSemicolonRadioButtonIsSelected() {
         AdminView adminView = new AdminView();
 
-        adminView.fileDelimiterSemicolon.setSelected(true);
 
         String result = adminView.getFileDelimiter();
         Assert.assertEquals(";", result);
